api/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
python-multipart